% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcade.R
\docType{data}
\name{arcade_df}
\alias{arcade_df}
\title{Data.frame of Arcade vector font information}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 366 rows and 5 columns.
}
\usage{
arcade_df
}
\description{
Data.frame of Arcade vector font information
}
\keyword{datasets}
