% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{create_text_df_single_row}
\alias{create_text_df_single_row}
\title{Create data.frame of glyph information for the given line of text.}
\usage{
create_text_df_single_row(text, dx = 2)
}
\arguments{
\item{text}{text string}

\item{dx}{character spacing in original units}
}
\value{
data.frame with coordinates for all the glyphs with characters offset
       appropriately.  \code{char_idx} is the index of the character within
       the given text string
}
\description{
Create data.frame of glyph information for the given line of text.
}
